requirements.txt

torch
torchvision
numpy
pandas
nltk
matplotlib
pillow
tokenizers
tqdm